+++ app/Http/Controllers/LocationController.php
@@ -12,6 +12,8 @@ class LocationController extends Controller
 {
   public function index()
   {
+    $this->authorize('viewAll', Location::class);
+
     return Inertia::render('Master/Locations/Index', [
       'locations' => Location::with('warehouse:id,name')->get()
     ]);
@@ -19,11 +21,14 @@ class LocationController extends Controller
 
   public function create()
   {
+    $this->authorize('create', Location::class);
     return Inertia::render('Master/Locations/Create');
   }
 
   public function store(Request $request)
   {
+    $this->authorize('create', Location::class);
+
     $validated = $request->validate([
       'name' => 'required|unique:locations,name',
       'type' => 'required',
@@ -41,6 +46,8 @@ class LocationController extends Controller
 
   public function show($id)
   {
+    $this->authorize('view', Location::class);
+
     return Inertia::render('Master/Locations/Create', [
       "location" => Location::with('warehouse:id,name')->where("id", $id)->first()
     ]);
@@ -48,6 +55,8 @@ class LocationController extends Controller
 
   public function update(Request $request, Location $location)
   {
+    $this->authorize('update', Location::class);
+
     $validated = $request->validate([
       'name' => 'required|unique:locations,name,' . $location->id,
       'type' => 'required',
@@ -63,6 +72,8 @@ class LocationController extends Controller
 
   public function destroy($id)
   {
+    $this->authorize('delete', Location::class);
+
     $ids = explode(',', $id);
     Location::whereIn('id', $ids)->delete();
     return Redirect::route('master.locations.index');
